(1) YOLO 대폭 수정     - YOLORegionLayer_device.cu : 인덱스 오류 수정     - YOLOPassthruLayer_device.cu : darknet, caffe의 reorg_layer와                                     동일하게 작동하도록 수정     - YOLOLossLayer.h : NMS의 Threshold값을 0.1 -> 0.3으로 수정     - yolo_train.json, yolo_test_live.json : 최신버전으로 변경     - 필요없는 네트워크 정의 파일 삭제
(2) NMS가 제대로 작동되지 않던 버그 수정
    - Worker.cpp : nms의 threshold 디폴트값과 관계없이 0으로 들어가는 문제 수정
    - ImageUtil.cpp : 이미지 하나에 동일 물체가 중복으로 감지 안되는 문제 수정

(3) caffemodel을 laonsill의 param으로 바꿔주는 파일 추가(YOLO 버전)
    - convert_caffe2laonsill_param.py

+++ LaonSill/src/core/Worker.cpp
@@ -719,7 +719,7 @@ void Worker::handleRunObjectDetectionNetworkWithInput(Job* job) {
         }
 
         // NMS 적용
-        int nmsThres = YOLO_DEFAULT_NMS_THRES;
+        float nmsThres = YOLO_DEFAULT_NMS_THRES;
         vector<yoloJobPack> nmsYoloPacks;
 
         for (int labelIdx = 0; labelIdx < classCount; labelIdx++) {
